bye bye make -j3, welcome makefile with libft dependance

+++ vm/load_champ2.c
@@ -6,7 +6,7 @@
 /*   By: fbabin <fbabin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/15 15:17:21 by fbabin            #+#    #+#             */
-/*   Updated: 2018/03/16 21:12:59 by arobion          ###   ########.fr       */
+/*   Updated: 2018/03/17 12:44:39 by fbabin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,10 +36,10 @@ int		check_champ_size(int fd, char *header, char *l)
 	while (read(fd, l, 1) > 0)
 	{
 		i++;
-		if (i > size + 10)
+		if (i > (int)size + 10)
 			break ;
 	}
-	if (i == size)
+	if (i == (int)size)
 		return (1);
 	ft_strdel(&header);
 	ft_strdel(&l);
